bugfix_ int used instead of uint64_t in main loop

+++ internal/dna_string_n.hpp
@@ -411,7 +411,7 @@ private:
 
 		bool res = true;
 
-		for(int i=0;i<size();++i){
+		for(uint64_t i=0;i<size();++i){
 
 			auto r = parallel_rank(i);
